app-indicator.c: Attempt to use an absolute path based on any application-provided icon location.
Slack creates a new temporary folder (in /tmp) any time it wants
a different icon.  Within slack, it adds the new folder to the icon
theme.  For our remote applet, it's better to just provide the
full path to that icon, instead of constantly changing the icon theme
search path.